Rename model of CAP implementation
- Rename from cap to alert
- Add version information to package path
