+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.0.0 (under development)
 This is the initial version setting up the base system
 
+- `CNG` Change name of CAP implementation to alert, add cap version to package name
 - `NEW` Add model of Common Alerting Protocol Version 1.2
 - `NEW` Add organization and category field to message payload
 - 'CNG' Raise quarkus version to 1.9.2
