+++ pom.xml
@@ -9,6 +9,7 @@
     <version>1.0.0-SNAPSHOT</version>
     <properties>
         <guava.version>30.0-jre</guava.version>
+        <javax-validation.version>2.0.1.Final</javax-validation.version>
         <lombok.version>1.18.16</lombok.version>
         <maven.compiler.version>3.8.1</maven.compiler.version>
         <maven.compiler.release>11</maven.compiler.release>
@@ -29,15 +30,16 @@
             <scope>provided</scope>
         </dependency>
         <dependency>
+            <groupId>javax.validation</groupId>
+            <artifactId>validation-api</artifactId>
+            <version>${javax-validation.version}</version>
+            <scope>provided</scope>
+        </dependency>
+        <dependency>
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
             <version>${lombok.version}</version>
             <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>javax.validation</groupId>
-            <artifactId>validation-api</artifactId>
-            <version>2.0.1.Final</version>
         </dependency>
     </dependencies>
