+++ src/main/java/de/dealog/common/emergency/cap/MsgType.java
@@ -1,4 +1,4 @@
-package de.dealog.cap.model;
+package de.dealog.common.emergency.cap;
 
 /**
  * The code denoting the nature of the alert message
