+++ src/main/java/de/dealog/common/emergency/cap/Scope.java
@@ -1,4 +1,4 @@
-package de.dealog.cap.model;
+package de.dealog.common.emergency.cap;
 
 /**
  * The code denoting the intended distribution of the alert message
