+++ src/main/java/de/dealog/common/emergency/cap/Urgency.java
@@ -1,4 +1,4 @@
-package de.dealog.cap.model;
+package de.dealog.common.emergency.cap;
 
 /**
  * The code denoting the urgency of the subject event of the alert message
